Backend: Resolved DELETE API errors

+++ app/api/delete_music/[id]/route.ts
@@ -1,18 +1,21 @@
 import Music from "@/libs/models/Music";
 import { connectMongoDB } from "@/libs/MongoConnect";
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 
 
 
-export async function DELETE (URLParams: any) {
+export async function DELETE (request: NextRequest, URLParams: any) {
     try {
-        const id = URLParams.params.id;
+        const { id }  = URLParams.params;
+        if (!id) {
+            throw new Error("ID is required");
+        }
         await connectMongoDB();
         await Music.findByIdAndDelete(id);
         return NextResponse.json({ msg: "Music Deleted Successfully!!"});
-    } catch (error) {
+    } catch (error: any) {
         return NextResponse.json({
-            error,
+            error: error.message,
             msg: "Some errors, Try Again!!",
         },
     { status: 400 });
